requirements.txt
fastapi
uvicorn
streamlit
requests
transformers
sentence-transformers
faiss-cpu
PyMuPDF  # for PDF ingestion
